Began working on optimising service pages
I'm combining agriculture, earthworks, and cartage into 1 file which will be called "services". I'll use variables and databases to swap out the image an text between them.

I also need to rename the page when it's finished and fix the links in the navbar

I organized the CSS so I could understand it better (because it was getting to big